requirements.txt
fastapi==0.115.12
fastapi[standard]
uvicorn==0.34.0
pydantic==2.11.2
python-dotenv==1.1.0
nest_asyncio==1.6.0
requests==2.32.3
beanie==1.29.0
pysqlite3==0.5.4
tqdm
matplotlib==3.9.2
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
pdfminer.six
pdfplumber==0.11.6
camelot-py==1.0.0
pdf2image==1.17.0
pdf2txt==0.7.14
torch
torchvision
transformers
openai==1.66.5
huggingface-hub==0.30.2
langchain==0.3.2
langchain-chroma==0.2.2
langchain-community==0.3.1
langchain-core==0.3.31
langchain-huggingface==0.1.0
langchain-openai==0.3.2
langchain-text-splitters==0.3.0
qwen_vl_utils==0.0.11
google-genai==1.12.1
google==3.0.0
google-auth==2.35.0
chromadb==0.6.3
motor==3.4.0
httpx==0.28.1
sqlalchemy==2.0.30
pytest
ipython